requirements.txt
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torchcodec==0.8.1
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@f065e402fc939758eabcfe182f34543258ac94ca
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
